refactor: split worker sentinel into v1/worker/sentinel for better plugin support and hardware adaptation
