+++ vllm/v1/fault_tolerance/__init__.py
@@ -3,11 +3,9 @@
 from .client_sentinel import ClientSentinel
 from .engine_core_sentinel import EngineCoreSentinel
 from .sentinel import BaseSentinel
-from .worker_sentinel import WorkerSentinel
 
 __all__ = [
     "BaseSentinel",
     "ClientSentinel",
     "EngineCoreSentinel",
-    "WorkerSentinel",
 ]
